+++ src/main/java/scanner/Main.java
@@ -3,8 +3,8 @@ package scanner;
 import lombok.extern.slf4j.Slf4j;
 import scanner.brute.BruteForceScanner;
 import scanner.brute.basic.BasicAuthScanner;
+import scanner.analyze.Resolver;
 import scanner.stat.ReportPrinter;
-import scanner.stat.ScreenStatGatherer;
 
 import java.net.InetSocketAddress;
 import java.util.Collections;
@@ -37,6 +37,7 @@ public class Main {
                 log.info("progress: {} {}/{}", ip, ++i, listSources.size());
                 bruteForceScanner.brute(ip, listPasswords.toArray(new String[0]));
             }
+            Resolver.run();
         }
 
         if (Preferences.check("-ba")) {