+++ src/main/java/scanner/analyze/AuthResolve.java
@@ -1,0 +1,10 @@
+package scanner.analyze;
+
+public class AuthResolve implements Resolve<String> {
+
+    @Override
+    public String resolve() {
+        return "admin:asdf1234";
+    }
+
+}