+++ src/main/java/scanner/analyze/PathResolve.java
@@ -1,0 +1,12 @@
+package scanner.analyze;
+
+import scanner.ffmpeg.FFmpegPath;
+
+public class PathResolve implements Resolve<FFmpegPath> {
+
+    @Override
+    public FFmpegPath resolve() {
+        return FFmpegPath.BASE;
+    }
+
+}