+++ src/main/java/scanner/analyze/ProblemTarget.java
@@ -1,0 +1,16 @@
+package scanner.analyze;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.RequiredArgsConstructor;
+
+@Data
+@EqualsAndHashCode
+@RequiredArgsConstructor
+public class ProblemTarget<T> {
+
+    private final String ip;
+    private final String credentials;
+    private final Resolve<T> resolve;
+
+}