+++ src/main/java/scanner/analyze/Resolve.java
@@ -1,0 +1,13 @@
+package scanner.analyze;
+
+/**
+ * Interface for providing a solution.
+ *
+ * @param <T> solution argument type
+ */
+public interface Resolve<T> {
+
+    T resolve();
+
+}
+