+++ src/main/java/scanner/ffmpeg/FFmpegExecutor.java
@@ -8,7 +8,7 @@ import scanner.stat.ScreenStatGatherer;
 import java.io.File;
 import java.util.concurrent.*;
 
-import static scanner.ffmpeg.FFmpegState.STANDARD;
+import static scanner.ffmpeg.FFmpegPath.STANDARD;
 
 /**
  * FFmpeg manipulation control class.
@@ -24,11 +24,11 @@ public class FFmpegExecutor {
      *
      * @param credentials login and password or null if absent
      * @param ip target address
+     * @param path rtsp stream path
      */
     @SneakyThrows
-    public void saveFrame(String credentials, String ip) {
-        ScreenStatGatherer.increment(ScreenStatEnum.ALL);
-        CompletableFuture.supplyAsync(() -> new FFmpegFrameSaver(ip, credentials, STANDARD).get())
+    public void saveFrame(String credentials, String ip, FFmpegPath path) {
+        CompletableFuture.supplyAsync(() -> new FFmpegFrameSaver(ip, credentials, path).get())
                 .thenAccept(x -> {
                     if (x.isAlive())
                         x.destroy();
@@ -39,6 +39,20 @@ public class FFmpegExecutor {
         File file = new File(String.format("result/screen/%s.jpg", ip));
         if (file.exists())
             ScreenStatGatherer.increment(ScreenStatEnum.SUCCESS);
+    }
+
+    /**
+     * Save one frame via FFmpeg and kill external process.
+     * After execution checked that frame download successfully and was saved.
+     * Use default stream path.
+     *
+     * @param credentials login and password or null if absent
+     * @param ip target address
+     */
+    @SneakyThrows
+    public void saveFrame(String credentials, String ip) {
+        ScreenStatGatherer.increment(ScreenStatEnum.ALL);
+        this.saveFrame(credentials, ip, STANDARD);
     }
 
 }