+++ src/main/java/scanner/stat/ScreenStatGatherer.java
@@ -1,9 +1,6 @@
 package scanner.stat;
 
-import java.time.LocalTime;
 import java.util.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import static scanner.stat.ScreenStatEnum.*;
 
@@ -41,15 +38,6 @@ public class ScreenStatGatherer {
         put(UNEXPECTED_ERROR, 0);
         put(OTHER, 0);
     }};
-
-    /**
-     * Gather info about rtsp server and streaming status from ffmpeg output.
-     *
-     * @param source ffmpeg output string
-     */
-    public static void gather(String source) {
-        findError(source);
-    }
 
     /**
      * Create full report by gather statistic data.
@@ -96,46 +84,6 @@ public class ScreenStatGatherer {
 
         screenStat.put(OTHER, other);
         screenStat.computeIfPresent(FAILURE, (x, y) -> y + other);
-    }
-
-    private static void findError(String source) {
-        Matcher mather;
-
-        mather = Pattern.compile("Could not find codec").matcher(source);
-        if (mather.find()) {
-            increment(NOT_FOUND_CODEC_ERROR);
-            return;
-        }
-        mather = Pattern.compile("Invalid data found").matcher(source);
-        if (mather.find()) {
-            increment(INVALID_DATA_FOUND);
-            return;
-        }
-        mather = Pattern.compile("401 Unauthorized").matcher(source);
-        if (mather.find()) {
-            increment(WRONG_AUTH_ERROR);
-            return;
-        }
-        mather = Pattern.compile("461 Unkown").matcher(source);
-        if (mather.find()) {
-            increment(PROTOCOL_NOT_SUPPORTED);
-            return;
-        }
-        mather = Pattern.compile("451 Parameter Not Understood").matcher(source);
-        if (mather.find()) {
-            increment(PARAMETER_NOT_UNDERSTOOD);
-            return;
-        }
-        mather = Pattern.compile("404 Stream Not Found").matcher(source);
-        if (mather.find()) {
-            increment(STREAM_NOT_FOUND);
-            return;
-        }
-        mather = Pattern.compile("400 Bad Request").matcher(source);
-        if (mather.find()) {
-            increment(BAD_REQUEST);
-            return;
-        }
     }
 
 }